requirements.txt
httpx>=0.27.0
playwright>=1.40.0
click>=8.1.0
rich>=13.0.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
anthropic>=0.70.0
gspread>=6.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
